requirements.txt
torch
torchvision
timm
segmentation-models-pytorch
scikit-learn
pandas
numpy
matplotlib
pillow
opencv-python
gradio
tqdm
